% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaseSyntheticFirms.R
\name{CreateBaseSyntheticFirms}
\alias{CreateBaseSyntheticFirms}
\title{Creates a set of simulated businesses}
\usage{
CreateBaseSyntheticFirms(L)
}
\arguments{
\item{L}{A list}
}
\value{
A list
}
\description{
\code{CreateBaseSyntheticFirms} creates a set of simulated businesses
}
