% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaseSyntheticFirms.R
\docType{data}
\name{CreateBaseSyntheticFirmsSpecifications}
\alias{CreateBaseSyntheticFirmsSpecifications}
\title{Specifications list for CreateBaseSyntheticFirms module}
\format{A list containing 6 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{NewInpTable}{new table to be created for datasets specified in the
 'Inp' specifications}
 \item{NewSetTable}{new table to be created for datasets specified in the
 'Set' specifications}
 \item{Inp}{scenario input data to be loaded into the datastore for this
 module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CreateBaseSyntheticFirms.R script.
}
\usage{
CreateBaseSyntheticFirmsSpecifications
}
\description{
A list containing specifications for the CreateBaseSyntheticFirms module.
}
\keyword{datasets}
