% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaseSyntheticFirms.R
\name{createBiz}
\alias{createBiz}
\title{Create biz table}
\usage{
createBiz(Biz_IsEs)
}
\arguments{
\item{Biz_IsEs}{A list of business by industry and employee size}
}
\value{
A list
}
\description{
\code{createBiz} Take the list of businesses by industry and employee size, reshape it
 to a long table with rows for each combination of industry and employee size, expand
 the list out so there is a row per business
}
