% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFutureSyntheticFirms.R
\name{CreateFutureSyntheticFirms}
\alias{CreateFutureSyntheticFirms}
\title{Creates a set of simulated businesses}
\usage{
CreateFutureSyntheticFirms(L)
}
\arguments{
\item{L}{A list}
}
\value{
A list
}
\description{
\code{CreateFutureSyntheticFirms} creates a set of simulated businesses
}
