% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFutureSyntheticFirms.R
\docType{data}
\name{CreateFutureSyntheticFirmsSpecifications}
\alias{CreateFutureSyntheticFirmsSpecifications}
\title{Specifications list for CreateFutureSyntheticFirms module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{NewSetTable}{new table to be created for datasets specified in the
 'Set' specifications}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CreateFutureSyntheticFirms.R script.
}
\usage{
CreateFutureSyntheticFirmsSpecifications
}
\description{
A list containing specifications for the CreateFutureSyntheticFirms module.
}
\keyword{datasets}
